requirements.txt
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
editor==1.6.6
idna==3.7
inquirer==3.2.4
python-dotenv==1.0.1
python-vlc==3.0.20123
readchar==4.1.0
requests==2.31.0
runs==1.2.2
six==1.16.0
urllib3==2.2.1
wcwidth==0.2.13
xmod==1.8.1
